+++ resources/views/user/orgDatasetView.blade.php
@@ -5,11 +5,9 @@
         @include('partials.alerts-bar')
         @include('partials.'. $root .'-nav-bar', ['view' => 'organisation'])
         @include('partials.org-nav-bar', ['view' => 'dataset', 'organisation' => $fromOrg])
-        <div class="row">
-            <div class="col-sm-3 col-xs-12 text-left p-l-none">
+        <div class="col-sm-3 col-xs-12">
             @include('partials.org-info', ['organisation' => $fromOrg])
-            </div>
-            @include('components.datasets.dataset_view', ['admin' => Auth::user()->is_admin])
         </div>
+        @include('components.datasets.dataset_view', ['admin' => Auth::user()->is_admin])
     </div>
 @endsection